move component example from page to mdx
