+++ src/docs/components/MdxComponents/index.ts
@@ -7,6 +7,7 @@ export { DocCollapsible } from "./DocCollapsible";
 export { DocPre } from "./DocPre";
 export { ComponentDemo } from "./ComponentDemo";
 export { ComponentSourceTabs } from "./ComponentSourceTabs";
+export { ComponentExamples } from "./ComponentExamples";
 export {
   getComponentSource,
   getComponentData,
