+++ src/docs/components/MdxComponents/utils.tsx
@@ -73,3 +73,54 @@ export const getComponentSource = async (componentType: string) => {
     })
   );
 };
+
+/**
+ * Maps over examples, translates them to html, and puts them together.
+ * @param type - A type of component. Same as `type` param of the page.
+ * @returns An object containing component title and array of component examples in React.
+ */
+export const getComponentExamples = async (componentType: string) => {
+  // Throws if component isn't registered
+  const component = getComponentData(componentType);
+
+  // Create directory path for component type
+  const typePath = join(process.cwd(), CONTENT_DIR, component.type);
+
+  // Read all the files in that dir
+  const files = readdirSync(typePath)
+    .filter((file) => file.endsWith(".tsx"))
+    .filter((file) => file !== "Demo.tsx");
+
+  // Initiate instance of highlighter
+  const highlighter = await getHighlighter();
+
+  const examples = await Promise.all(
+    files.map(async (file) => {
+      const id = file.replace(/.tsx/, "");
+
+      const data = readFileSync(join(typePath, file), "utf8");
+      const Component = (
+        await import(`src/docs/examples/${component.type}/${id}.tsx`)
+      ).default;
+
+      const html = render(<Component />, { pretty: true });
+      const plainText = render(<Component />, { plainText: true });
+
+      const source = await highlight(highlighter, data);
+      const markup = await highlight(highlighter, html, "html");
+      return {
+        id,
+        html,
+        source,
+        markup,
+        plainText,
+        type: component.type,
+      };
+    })
+  );
+
+  return {
+    title: component.title,
+    examples,
+  };
+};
