+++ src/docs/content/components/badges.mdx
@@ -5,6 +5,7 @@ Badge like component
 import {
   ComponentDemo,
   ComponentSourceTabs,
+  ComponentExamples,
 } from "src/docs/components/MdxComponents";
 
 <ComponentDemo type="badges" />
@@ -15,3 +16,7 @@ You can use components from this library by copying them to your project.
 It's recommended to copy source code into folder aliased with **'@mailingui/components'**
 
 <ComponentSourceTabs type="badges" />
+
+## Examples
+
+<ComponentExamples type="badges" />
