+++ src/docs/content/components/buttons.mdx
@@ -5,6 +5,7 @@ Drive user action with clickable buttons
 import {
   ComponentDemo,
   ComponentSourceTabs,
+  ComponentExamples,
 } from "src/docs/components/MdxComponents";
 
 <ComponentDemo type="buttons" />
@@ -15,3 +16,7 @@ You can use components from this library by copying them to your project.
 It's recommended to copy source code into folder aliased with **'@mailingui/components'**
 
 <ComponentSourceTabs type="buttons" />
+
+## Examples
+
+<ComponentExamples type="buttons" />
