+++ src/docs/content/components/dividers.mdx
@@ -5,6 +5,7 @@ Create visual separation for a clean, professional look
 import {
   ComponentDemo,
   ComponentSourceTabs,
+  ComponentExamples,
 } from "src/docs/components/MdxComponents";
 
 <ComponentDemo type="dividers" />
@@ -15,3 +16,7 @@ You can use components from this library by copying them to your project.
 It's recommended to copy source code into folder aliased with **'@mailingui/components'**
 
 <ComponentSourceTabs type="dividers" />
+
+## Examples
+
+<ComponentExamples type="dividers" />
