+++ src/docs/content/components/emojis.mdx
@@ -5,6 +5,7 @@ Add personality and emotion to your messages
 import {
   ComponentDemo,
   ComponentSourceTabs,
+  ComponentExamples,
 } from "src/docs/components/MdxComponents";
 
 <ComponentDemo type="emojis" />
@@ -15,3 +16,7 @@ You can use components from this library by copying them to your project.
 It's recommended to copy source code into folder aliased with **'@mailingui/components'**
 
 <ComponentSourceTabs type="emojis" />
+
+## Examples
+
+<ComponentExamples type="emojis" />
