+++ src/docs/content/components/hero-sections.mdx
@@ -5,6 +5,7 @@ Beautiful headers to showcase your message
 import {
   ComponentDemo,
   ComponentSourceTabs,
+  ComponentExamples,
 } from "src/docs/components/MdxComponents";
 
 <ComponentDemo type="hero-sections" />
@@ -15,3 +16,7 @@ You can use components from this library by copying them to your project.
 It's recommended to copy source code into folder aliased with **'@mailingui/components'**
 
 <ComponentSourceTabs type="hero-sections" />
+
+## Examples
+
+<ComponentExamples type="hero-sections" />
