+++ src/docs/content/components/lists.mdx
@@ -5,6 +5,7 @@ All lists components
 import {
   ComponentDemo,
   ComponentSourceTabs,
+  ComponentExamples,
 } from "src/docs/components/MdxComponents";
 
 <ComponentDemo type="lists" />
@@ -16,3 +17,7 @@ It's recommended to copy source code into folder aliased with **'@mailingui/comp
 For **BulletList** support you have to copy that one too.
 
 <ComponentSourceTabs type="lists" />
+
+## Examples
+
+<ComponentExamples type="lists" />
