+++ src/docs/content/components/social-icons.mdx
@@ -5,6 +5,7 @@ Increase engagement with clickable social icons
 import {
   ComponentDemo,
   ComponentSourceTabs,
+  ComponentExamples,
 } from "src/docs/components/MdxComponents";
 
 <ComponentDemo type="social-icons" />
@@ -15,3 +16,7 @@ You can use components from this library by copying them to your project.
 It's recommended to copy source code into folder aliased with **'@mailingui/components'**
 
 <ComponentSourceTabs type="social-icons" />
+
+## Examples
+
+<ComponentExamples type="social-icons" />
