+++ src/docs/content/components/texts.mdx
@@ -5,6 +5,7 @@ Craft compelling content with easy-to-read paragraphs
 import {
   ComponentDemo,
   ComponentSourceTabs,
+  ComponentExamples,
 } from "src/docs/components/MdxComponents";
 
 <ComponentDemo type="texts" />
@@ -15,3 +16,7 @@ You can use components from this library by copying them to your project.
 It's recommended to copy source code into folder aliased with **'@mailingui/components'**
 
 <ComponentSourceTabs type="texts" />
+
+## Examples
+
+<ComponentExamples type="texts" />
